Custom-Tool-Example/requirements.txt
langchain
langchain-community
langchain-experimental
boto3
python-dotenv
spotipy 